style: move IntModN to modular.rs
