+++ toolchain/riscv/axvm/src/intrinsics/mod.rs
@@ -9,8 +9,8 @@ pub use hash::*;
 #[cfg(target_os = "zkvm")]
 pub use io::*;
 
-mod modulus;
-pub use modulus::*;
+mod modular;
+pub use modular::*;
 
 mod utils;
 use utils::*;
