--- toolchain/riscv/axvm/src/intrinsics/modulus/mod.rs
@@ -1,2 +1,0 @@
-mod intmodn;
-pub use intmodn::*;
